requirement.txt
appdirs==1.4.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
distlib==0.3.0
distro==1.4.0
efficientnet-pytorch==0.7.0
html5lib==1.0.1
idna==2.8
imageio==2.9.0
ipaddr==2.2.0
kiwisolver==1.3.1
lockfile==0.12.2
matplotlib==3.3.3
msgpack==0.6.2
numpy==1.19.4
packaging==20.3
pep517==0.8.2
Pillow==8.0.1
progress==1.5
protobuf==3.14.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
scipy==1.5.4
six==1.14.0
tensorboardX==2.1
torch==1.7.1+cu101
torchaudio==0.7.2
torchvision==0.8.2+cu101
typing-extensions==3.7.4.3
urllib3==1.25.8
webencodings==0.5.1
